Internals

Index

BayesJ.cosmo

BayesJ.μ

BayesJ.μ_e

BayesJ.Dataset

BayesJ.Ellipse

BayesJ.LinearInterpolator

BayesJ.ObservationError

BayesJ.Prior

BayesJ.XS_Mekal

BayesJ.apply_response_function

BayesJ.bin_events

BayesJ.bin_events

BayesJ.bin_events

BayesJ.call_mekal

BayesJ.cluster

BayesJ.find_blobs_2D

BayesJ.fit_ellipse

BayesJ.getellipsepoints

BayesJ.load_data

BayesJ.load_data

BayesJ.load_data

BayesJ.load_events_from_fits

BayesJ.load_mask

BayesJ.load_mask

BayesJ.load_response

BayesJ.load_response

BayesJ.load_response

BayesJ.log_factorial

BayesJ.log_likelihood

BayesJ.make_cube_transform

BayesJ.make_observation

BayesJ.make_observation

BayesJ.pixel_offset

BayesJ.surface_brightness

BayesJ.test_point

BayesJ.transform

BayesJ.ρ_crit

Private Interface

BayesJ.cosmo — Constant

Cosmology assumed by BayesX

Defaults to:

h=0.69,

Neff=3.04,

OmegaK=0,

OmegaM=0.29,

OmegaR=nothing,

Tcmb=2.7255,

w0=-1,

wa=0

BayesJ.μ — Constant

Mean mass per gas particle = 0.6mₚ

BayesJ.μ_e — Constant

Mean gas mass per electron = 1.14mₚ

BayesJ.Dataset — Type

Abstract type to wrap various specific formats for source data. By doing so we can make use of multiple dispatch to automatically select the appropriate loading functions for data.

BayesJ.Ellipse — Type

Ellipse(x, y, r1, r2, θ)

An ellipse centered at (x, y) with radii r1 and r2 and rotation θ.

r1 and r2 are the width and height respectively with zero rotation and θ is defined as rotation counterclockwise from the x-axis in radians.

BayesJ.LinearInterpolator — Type

LinearInterpolator(x, y)

Construct a linear interpolator for the given data.

Boundary behaviour

If x is below the minimum value of x in the data it is rounded up the the minimum. This results in a flat line below the minimum value of x. In the cluster context this means the core properties may be constant below a certain radius. To avoid this simply fix the smallest value of x` at zero.

If x is above the maximum value of x in the data the interpolator returns zero. This is because the density and temperature are assumed to be zero outside the cluster. Fix the density and temperature at your largest radius to zero if you wish to avoid a discontinuity here.

BayesJ.ObservationError — Type

ObservationError(likelihood)

This wraps a fallback likelihood value so it can be passed up the chain on an invalid prior combination.

BayesJ.Prior — Type

Abstract supertype for priors. Should implement:

A transform(prior, x) function that transforms a value x on the unit range to a value on the distribution represented by the prior.

A field name<:AbstractString

BayesJ.XS_Mekal — Type

Mekal model using SpectralFitting.jl framework. Kept around so we can borrow the model data downloading functions.

BayesJ.apply_response_function — Method

apply_response_function(counts_per_bin::Vector, response::Matrix, exposure_time::Unitful.Time)::Vector{Float64}

Applies the response function $RSP(PI, E) = RMF(PI, E) ∘ ARF(E)$ to all energy bins and returns adjusted counts per bin.

\[\begin{aligned} C(PI) &= T \int RMF(PI, E) ⋅ ARF(E) ⋅ S(E) ⋅ dE \\ &≈ T \sum_{j} R_{ij} A_{j} S_{j} \end{aligned}\]

with C(PI) is the observed counts in a detector channel PI, T is the observation time, ARF(E) is the effective area of the instrument and RMF(E, PI) [we use RMF(PI, E)] is the unitless response matrix. S(E) is the source flux, in [photon count]/[time]/[area]/[energy]. (Handbook of X-ray Astronomy Chapter 2011, Chapter 5 pg 87, by K. Arnaud, R. Smith and A. Siemiginowska)

This function takes the combined RMF and ARF as the response function. This is to recalculating it on every call. Some people format the RMF as RMF(E, PI), others as RMF(PI, E). This latter convention is used by CIAO and by us.

BayesJ.bin_events — Method

bin_events(::Dataset, events, ...)

Placeholder function, returns events matrix without modification

BayesJ.bin_events — Method

bin_events(::FITSData, events, energy_range, x_edges, y_edges)

Take a table of events in the format (x y channel), trim it by energy and bin it spatially. Returns an array of counts per bin with dimensions (channel, x, y).

BayesJ.bin_events — Method

bin_events(::PlaintextData, events, ...)

Placeholder function, returns events matrix without modification

BayesJ.call_mekal — Method

call_mekal(n_energy_bins, abundances, min_energy, max_energy, bin_sizes, temperature, nH)

Given a unitful range of energy and unitless temperature (keV) and hydrogen density in the source (cm^-3) calls MEKAL to calculate the volume emissivity of the source in the source frame.

Result is in photons/m^3/s/bin.

Metal abundances are set by abundances. They include H, He, C, N, O, Ne, Na, Mg, Al, Si, S, Ar, Ca, Fe and Ni and are specified with respect to solar values. MEKAL uses the solar values from (Anders and Grevesse, 1989).

Due to performance concerns I recommend calling prepare_model_mekal instead to generate an interpolation over the model. The functions returned by that wrapper also implicitly include absorption, which this function does not.

Implementation Details

The XSPEC and MEKAL documentation is at times short on detail. Here's what I've figured out, mostly by inspecting the models in the XSPEC source code. The MEKA model has similar code and more extensive comments that MEKAL that are very helpful.

The values calculated by MEKAL are multiplied by a constant of 2.53325e-3. Inspecting the MEKA comments we find the following equation for this constant which we will term C

\[C = √2 ⋅ h^2 ⋅ α^3 / ((3πm_e)^1.5 ⋅ √(1000e) ⋅ π ⋅ (1pc)^2 ⋅ 10^4)\]

and the instruction to use SI units. h appears to be Planck's constant (unreduced) and α the fine-structure constant. Calculating this gives C=2.533242122484874e-59 kg^1/2 m^2 A^-1/2 s^-5/2.

MEKAL adjusts emissions by what it terms cem which is the emitting volume divided by the square of distance to the volume. This follows naturally from the relationship between surface emissivity σ and volume emissivity ϵ.

\[ σ = ϵV / (4πD)^2\]

MEKAL uses units of 1e50cm^3 / 1pc^2 for cem. If we multiply by 10^50cm^3 we get 2.533242122484874e-13 J^2 cm s^2 C^-1/2 kg^-3/2. As dividing by 1pc^2 would change more than just the magnitude of the value I assume it is folded in as the 1pc^2 in the original equation. Furthermore I assume that remaining difference rests in the details of units and implementation.

MEKAL states it outputs in phot/cm^2/s/keV. MEKA states the same but has an optional switch that instead outputs in phot/m^3/s/keV. It does this by replacing C with D=3.03103E-9 which it states is derived so that

\[C = D * 10^40 / (4π * 1pc^2)\]

I have verified this gives results matching that of doing the conversion manually (by dividing out 1e50cm^3 and multiplying by 1pc^2).

Attempts were made to invoke MEKAL through XSPEC's wrapper functions but I found the modifications they introduced were troublesome to work out, though some were necessary and are reimplemented in surface_brightness. MEKAL's operations are more clearly physically motivated.

BayesJ.cluster — Function

Cluster pixels from blobs

BayesJ.find_blobs_2D — Method

Find blobs in a 2D matrix

BayesJ.fit_ellipse — Method

Fit an ellipse to a set of pixels

BayesJ.getellipsepoints — Method

Get points from the boundry of the ellipse

BayesJ.load_data — Method

load_data(data)

Load events data from a given dataset, returning pairs of events and exposure times.

The first pair includes all observed events, the second is the background observation.

BayesJ.load_data — Method

load_data(data::FITSData)

Load data from FITS events files for the observation and background.

Exposure times are taken from the LIVETIME value in the header.

BayesJ.load_data — Method

load_data(data::PlaintextData)

Load data from plaintext files as generated by existing Bayes-X binning scripts.

Data is reshaped to the given size so should be provided in column-major order or in a matrix of correct size.

BayesJ.load_events_from_fits — Method

load_events_from_fits(path)

Loads events from a single fits file for further processing.

BayesJ.load_mask — Method

load_mask(path, x_edges, y_edges)

Takes a mask file as used by DS9 and x and y edges to bin on. This should match the edges used to bin the data.

BayesJ.load_mask — Method

load_mask(data)

The matrix is true for masked pixels.

BayesJ.load_response — Method

load_response(data, min_energy, max_energy)

Load the RMF and ARF for an observation, trimmed for the desired energy range

BayesJ.load_response — Method

load_response(data::FITSData, min_energy::Unitful.Energy, max_energy::Unitful.Energy)

Load response function for specified energy range, adjusting it to align with channel edges.

Returns response matrix, energy bins and included channels. Assumes the RMF and ARF are formatted according to the HEASARC OGIP Calibration Memo CAL/GEN/92-002

BayesJ.load_response — Method

load_response(data::PlaintextData)

Load response matrix from plaintext files as generated by existing Bayes-X binning scripts.

Data is reshaped to the given size so should be provided in column-major order or as a matrix of correct size.

BayesJ.log_factorial — Method

log_factorial(n)

Finds $\ln{n!}$ the natural logarithm of the factorial of n.

As Γ(n) = (n-1)! we can use SpecialFunctions.jl loggamma function to quickly calculate the log factorial.

It is intended to be broadcast across all values of the data array.

BayesJ.log_likelihood — Method

log_likelihood(observed, observed_background, predicted, predicted_background, observed_log_factorial)

Calculate the log-likelihood of the prediction given an observation.

The observed and predicted arrays include background events.

The constant likelihood includes all parts of the likelihood that remain constant across observations. We require it to be supplied to improve performance - no need to calculate it every time. It also includes the log factorial components calculated as ln(observed) + ln(observed_background). As we haven't implemented support for background fitting on the user input side of things this includes the background.

BayesJ.make_cube_transform — Method

make_cube_transform(priors::Prior...)::NTuple{2, Function}

Turn a sequence of prior transform functions into a transform function transform_cube that operates on the hypercube generated by multinest, stripping delta functions.

It also returns a function reconstruct_args take takes the output of the first function and returns it with the values of the delta priors inserted in appropriate postions. This is because Ultranest can run with delta priors but outputs error warnings in the process.

Returns transform_cube and reconstruct_args as a tuple.

BayesJ.make_observation — Method

make_observation(temperature, density, z, shape, pixel_edge_angle, emission_model, exposure_time, response_function, centre, centre_radius, mask=nothing)

Unitless wrapper for make_observation

BayesJ.make_observation — Method

make_observation(temperature, density, z, shape, pixel_edge_angle, emission_model, exposure_time, response_function, centre, centre_radius, mask=nothing, limit=10u"Mpc")

Generate an image of the cluster given functions for the radial profile of gas temperature and electron density and assorted observational parameters.

shape describes the size of the source number as a number of spatial bins in each dimension.-

The pixel edge angle describes the angular size observed by a single pixel in units such as arcseconds. This area is assumed to be square with the edge angle giving the side length.

The emission model should be a function compatible with the requirements of the surface_brightness function, which it will be passed to.

The response function includes both the RMF and ARF, as described in apply_response_function.

The limit is passed through to surface_brightness

BayesJ.pixel_offset — Method

pixel_offset(ij, array_centre_indices, centre_offset_pixels)

Calculate the distance from the centre of the cluster to the pixel at indices ij.

Result in pixels.

BayesJ.surface_brightness — Method

surface_brightness(projected_radius, temperature, density, z, limit, model, pixel_edge_angle)

Calculate the observed surface_brightness at some projected radius on the sky.

The temperature and density functions should take the distance from origin (cluster centre) and return the appropriate value, with units.

The model should take these functions as input and return a vector of volume emissivities, where the vector element correspond with the energy bins used to generate the mode. It should include the effects of redshift on the bins and time dilation on the count rate.

The limit controls limits on LOS radius during integration. Testing suggests that Inf (default) is faster than finite values >1. However finite values may avoid problems with models that are badly behaved at large radii.

The pixel edge angle is the angular size of a single pixel.

The result returned is the expected count rate per unit observing area, as a vector corresponding to the energy bins as used in the model.

BayesJ.test_point — Method

test_point(e::Ellipse, x, y; atol=0)

Tests whether a point is within an ellipse with tolerance atol.

The equation for an ellipse in Cartesian coordinates is of the form f(x,y) = 1. We consider any point (x0, y0) such that f(x0, y0) <= 1 + atol as being within the ellipse.

BayesJ.transform — Method

transform(prior, x)

Transforms a value x on the unit range to a value on the distribution represented by the prior.

BayesJ.ρ_crit — Method

p_crit(z)

Calculate the critical density at some redshift z.