Internals

Contents

Internals

Contents

Index

Private Interface

Cluster Models

Constants

Emissions Modelling

IO

Likelihood

Utilities

Index

BayesJ.cosmo

BayesJ.μ

BayesJ.μ_e

BayesJ.Dataset

BayesJ.Prior

BayesJ.XS_Mekal

BayesJ.Model_NFW_GNFW

BayesJ.apply_response_function

BayesJ.bin_events

BayesJ.bin_events

BayesJ.call_mekal

BayesJ.load_data

BayesJ.load_data

BayesJ.load_data

BayesJ.load_events_from_fits

BayesJ.load_response

BayesJ.load_response

BayesJ.log_factorial

BayesJ.log_likelihood

BayesJ.make_cube_transform

BayesJ.prepare_model_mekal

BayesJ.surface_brightness

BayesJ.transform

BayesJ.ρ_crit

Private Interface

Cluster Models

BayesJ.Model_NFW_GNFW — Method

Model_NFW_GNFW_GNFW(MT_200, fg_200, α, β, γ, c_500_GNFW, z, shape, pixel_edge_angle, emission_model, exposure_time, response_function)

Calculate predicted counts using a physical model based NFW-GNFW profiles as described in Olamaie 2012.

The first six parameters are model priors. z is redshift and shape describes the size of the source number as a number of spatial bins in each dimension. The pixel edge angle describes the angular size observed by a single pixel in units such as arcseconds. This area is assumed to be square with the edge angle giving the side length. The emission model should be a function compatible with the requirements of the surface_brightness function, which it will be passed to. The response function includes both the RMF and ARF, as described in apply_response_function.

BayesJ.ρ_crit — Method

p_crit(z)

Calculate the critical density at some redshift z.

Constants

BayesJ.cosmo — Constant

Cosmology assumed by BayesX

Defaults to:

h=0.69,

Neff=3.04,

OmegaK=0,

OmegaM=0.29,

OmegaR=nothing,

Tcmb=2.7255,

w0=-1,

wa=0

BayesJ.μ — Constant

Mean mass per gas particle = 0.6mₚ

BayesJ.μ_e — Constant

Mean gas mass per electron = 1.14mₚ

Emissions Modelling

BayesJ.apply_response_function — Method

apply_response_function(counts_per_bin::Vector, response::Matrix, exposure_time::Unitful.Time)::Vector{Float64}

Applies the response function $RSP(PI, E) = RMF(PI, E) ∘ ARF(E)$ to all energy bins and returns adjusted counts per bin.

\[\begin{aligned} C(PI) &= T \int RMF(PI, E) ⋅ ARF(E) ⋅ S(E) ⋅ dE \\ &≈ T \sum_{j} R_{ij} A{j} S{j} \end{aligned}\]

with C(PI) is the observed counts in a detector channel PI, T is the observation time, ARF(E) is the effective area of the instrument and RMF(E, PI) is the unitless response matrix. (Handbook of X-ray Astronomy Chapter 2011, Chapter 5 pg 87, by K. Arnaud, R. Smith and A. Siemiginowska)

This function takes the combined RMF and ARF as the response function. This is to recalculating it on every call. Some people format the RMF as RMF(PI, E). This convention is used by CIAO, for example.

BayesJ.surface_brightness — Method

surface_brightness(projected_radius, temperature, density, z, limit, model, pixel_edge_length)

Calculate the observed surface_brightness at some projected radius on the sky.

The temperature and density functions should take the true distance from origin (cluster centre) and return the appropriate value, with units. The limit controls limits on LOS radius during integration. Theoretically it should be ±infinity, but we may approximate it as a finite value. The pixel edge length is the distance observed on the sky, determined from the pixeledgeangle at source redshift.

BayesJ.XS_Mekal — Type

Mekal model using SpectralFitting.jl framework. Kept around so we can borrow the model data downloading functions.

BayesJ.call_mekal — Method

call_mekal(energy_range, temperature, nH)

Given a unitful range of energy and unitless temperature (keV) and hydrogen density in the source (cm^-3) calls MEKAL to calculate the volume emissivity of the source in the source frame.

Due to performance concerns I recommend calling prepare_model_mekal instead to generate an interpolation over the model. The interpolation also implicitly includes absorption, which this function does not.

Implementation Details

The XSPEC and MEKAL documentation is at times short on detail. Here's what I've figured out, mostly by inspecting the models in the XSPEC source code. The MEKA model has similar code and more extensive comments that MEKAL that are very helpful.

The values calculated by MEKAL are multiplied by a constant of 2.53325e-3. Inspecting the MEKA comments we find the following equation for this constant which we will term C

\[C = √2 ⋅ h^2 ⋅ α^3 / ((3πm_e)^1.5 ⋅ √(1000e) ⋅ π ⋅ (1pc)^2 ⋅ 10^4)\]

and the instruction to use SI units. h appears to be Planck's constant (unreduced) and α the fine-structure constant. Calculating this gives C=2.533242122484874e-59 kg^1/2 m^2 A^-1/2 s^-5/2.

MEKAL adjusts emissions by what it terms cem which is the emitting volume divided by the square of distance to the volume. This follows naturally from the relationship between surface emissivity σ and volume emissivity ϵ.

\[ σ = ϵV / (4πD)^2\]

MEKAL uses units of 1e50cm^3 / 1pc^2 for cem. If we multiply by 10^50cm^3 we get 2.533242122484874e-13 J^2 cm s^2 C^-1/2 kg^-3/2. As dividing by 1pc^2 would change more than just the magnitude of the value I assume it is folded in as the 1pc^2 in the original equation. Furthermore I assume that remaining difference rests in the details of units and implementation.

MEKAL states it outputs in phot/cm^2/s/keV. MEKA states the same but has an optional switch that instead outputs in phot/m^3/s/keV. It does this by replacing C with D=3.03103E-9 which it states is derived so that

\[C = D * 10^40 / (4π * 1pc^2)\]

I have verified this gives results matching that of doing the conversion manually (by dividing out 1e50cm^3 and multiplying by 1pc^2).

Attempts were made to invoke MEKAL through XSPEC's wrapper functions but I found the modifications they introduced were troublesome to work out, though some were necessary and are reimplemented in surface_brightness. MEKAL's operations are more clearly physically motivated.

BayesJ.prepare_model_mekal — Method

prepare_model_mekal(nHcol, energy_bins, z; temperatures, hydrogen_densities)

Create an interpolated alias to the mekal model with photoelectric absorption over specified parameter ranges.

The interpolation object (T::Unitful.Energy, nH::Unitful.𝐋^-3) -> volume emissivity returned gives the volume emissivity per bin of a source of the specified temperature and hydrogen number density with the effects of absorption from passage through an area with a hydrogen column density equal to that specified in the prepare_model_mekal call.

Applies redshift and time dilation.

Interpolation has a significant performance improvement over calling the model directly. No redshift is currently applied to energy bins - they should be assumed to be in the source frame. It may be wise to adjust this but then surface_brightness will need adjustment so it doesn't apply the correction twice.

IO

BayesJ.Dataset — Type

Abstract type to wrap various specific formats for source data. By doing so we can make use of multiple dispatch to automatically select the appropriate loading functions for data.

BayesJ.bin_events — Method

bin_events(::FITSData, events, energy_range, x_edges, y_edges)

Take a table of events in the format (x y channel energy), trim it by energy and bin it spatially. Returns an array of counts per bin with dimensions (channel, x, y).

BayesJ.bin_events — Method

bin_events(::PlaintextData, events, ...)

Placeholder function, returns events matrix without modification

BayesJ.load_data — Method

load_data(data)

Load events data from a given dataset, returning pairs of events and exposure times.

The first pair includes all observed events, the second is the background observation.

BayesJ.load_data — Method

load_data(data::FITSData)

Load data from FITS events files for the observation and background.

Exposure times are taken from the LIVETIME value in the header.

BayesJ.load_data — Method

load_data(data::PlaintextData)

Load data from plaintext files as generated by existing Bayes-X binning scripts.

Data is reshaped to the given size so should be provided in column-major order or in a matrix of correct size.

BayesJ.load_events_from_fits — Method

load_events_from_fits(path)

Loads events from a single fits file for further processing.

BayesJ.load_response — Method

load_response(data, energy_range)

Load the RMF and ARF for an observation, trimmed for the desired energy range

BayesJ.load_response — Method

load_response(data::PlaintextData, _)

Load response matrix from plaintext files as generated by existing Bayes-X binning scripts.

Data is reshaped to the given size so should be provided in column-major order or as a matrix of correct size.

Likelihood

BayesJ.Prior — Type

Abstract supertype for priors. Should implement a transform(prior, x) function that Transforms a value x on the unit range to a value on the distribution represented by the prior.

BayesJ.log_factorial — Method

log_factorial(n)

Finds the natural logarithm of the factorial of n.

n rapidly gets to large to quickly and directly calculate the factorial so we exploit logarithm rules to expand it out to a series of sums.

It is intended to be broadcast across all values of the data array.

BayesJ.log_likelihood — Method

log_likelihood(observed, observed_background, predicted, predicted_background, observed_log_factorial)

Calculate the log-likelihood of the prediction given an observation.

The observed and predicted arrays include background events. Log factorial is calculated as ln(observed) + ln(observed_background). We require it to be supplied to improve performance - no need to calculate it every time.

We assume the predicted_background is scaled to the same exposure time as the observed background.

BayesJ.make_cube_transform — Method

make_cube_transform(priors::Prior...)

Turn a sequence of prior transform functions intoNorma a transform function that operates on the hypercube generated by multinest.

BayesJ.transform — Method

transform(prior, x)

Transforms a value x on the unit range to a value on the distribution represented by the prior.

Utilities