Public Documentation

Index

BayesJ.DeltaPrior

BayesJ.DependentLogUniformPrior

BayesJ.DependentUniformPrior

BayesJ.FITSData

BayesJ.GenericPrior

BayesJ.LogUniformPrior

BayesJ.NormalPrior

BayesJ.PlaintextData

BayesJ.PriorError

BayesJ.UniformPrior

BayesJ.Model_Constant

BayesJ.Model_Constant

BayesJ.Model_Einasto

BayesJ.Model_Einasto

BayesJ.Model_NFW

BayesJ.Model_NFW

BayesJ.Model_Piecewise

BayesJ.Model_Vikhlinin2006

BayesJ.Model_Vikhlinin2006

BayesJ.prepare_model_mekal

BayesJ.prepare_model_mekal_interpolation

BayesJ.priorcheck

BayesJ.sample

BayesJ.sample

Public Interface

BayesJ.DeltaPrior — Type

DeltaPrior(name::AbstractString, value::Number)

A delta prior that always returns a constant value.

BayesJ.DependentLogUniformPrior — Type

DependentLogUniformPrior(name::AbstractString, depends_on::AbstractString, range::Number)

Generates a dependent log uniform prior.

The dependent prior has a minimum value equal to that of the parent prior and a maximum value equal to the parent prior plus the range. By specifying a negative range, the dependent prior can be made to have a maximum value less than the parent prior.

BayesJ.DependentUniformPrior — Type

DependentUniformPrior(name::AbstractString, depends_on::AbstractString, range::Number)

Generates a dependent uniform prior.

The dependent prior has a minimum value equal to that of the parent prior and a maximum value equal to the parent prior plus the range. By specifying a negative range, the dependent prior can be made to have a maximum value less than the parent prior.

BayesJ.FITSData — Type

FITSData(observation::AbstractString, background::AbstractString, arf::AbstractString, rmf::AbstractString, pixel_edge_angle::DimensionfulAngles.Angle)

Store a collection of observation data from fits files.

Takes paths to the files as inputs.

BayesJ.GenericPrior — Type

GenericPrior(name::AbstractString, dist::Distributions.UnivariateDistribution)

A generic prior that draws from the specified distribution.

See Distributions.jl for a list of avaliable distributions.

BayesJ.LogUniformPrior — Type

LogUniformPrior(name::AbstractString, min::Number, max::Number)

A log uniform prior that draws from a distribution between min and max whose base 10 logarithm is uniformly distributed.

BayesJ.NormalPrior — Type

NormalPrior(name::AbstractString, mean::Number, σ::Number)

A normal prior that draws from a Gaussian/normal distribution with mean and standard deviation σ.

BayesJ.PlaintextData — Type

PlaintextData(observation::AbstractString, background::AbstractString, arf::AbstractString, rmf::AbstractString, shape::NTuple{3,Int}, rmf_shape::NTuple{2,Int}, obs_exposure_time::Unitful.Time, bg_exposure_time::Unitful.Time, pixel_edge_angle::DimensionfulAngles.Angle, rmf_unit::Unitful.Area)

Store preprocessed data from plaintext files with whitespace delimiters.

Events and response files provided as paths. This data is assumed to have the appropriate binning and energy ranges already applied. The shape of the event arrays should be (channels, x, y) and the rmf_shape should be (channels, bins).

BayesJ.PriorError — Type

PriorError(likelihood)

This wraps a fallback likelihood value so it can be passed up the chain on an invalid prior combination.

BayesJ.UniformPrior — Type

UniformPrior(name::AbstractString, min::Number, max::Number)

A uniform prior that draws from a uniform distribution between min and max.

BayesJ.Model_Constant — Method

Model_Constant(T::Unitful.Energy, ρ::Unitful.Density; kwargs...)

Unitless wrapper for Model_Constant

Radius is in Mpc, temperature is in keV, density is in g/cm^3

BayesJ.Model_Constant — Method

Model_Constant(radius::Unitful.Length, T::Unitful.Energy, ρ::Unitful.Density; kwargs...)

Constant temperature and density model in sphere of given radius.

Intended for testing purposes.

BayesJ.Model_Einasto — Method

Model_Einasto(MT_Δ::Real, fg_Δ, n, α, β, γ, c_Δ_GNFW; z, Δ=500)

Unitless wrapper for Model_Einasto. Mass is in solar masses.

BayesJ.Model_Einasto — Method

Model_Einasto(MT_Δ::Unitful.Mass, fg_Δ, c_Δ_dm, n, α, β, γ, c_Δ_GNFW; z, Δ=500)

Generate a cluster profile based on the Einasto mass density profile and the GNFW gas pressure profile.

See Appendix A of (Olamaie et al., 2015) for a derivation of the model based on these profiles.

Returns functions for gas temperature and gas mass density as a function of radius.

If 1/n>2 then the gas temperature starts increasing to physically improbable levels (orders exceeding 10¹¹ keV).

BayesJ.Model_NFW — Method

Model_NFW(MT_Δ::Real, fg_Δ, α, β, γ, c_Δ_GNFW; z, Δ=500)

Unitless wrapper for Model_NFW. Mass is in solar masses.

BayesJ.Model_NFW — Method

Model_NFW(MT_Δ::Unitful.Mass, fg_Δ, c_Δ_dm, α, β, γ, c_Δ_GNFW; z, Δ=500)

Generate a cluster profile based on the NFW mass density and GNFW gas density profiles.

Uses the model from (Olamaie et al., 2012), which is based on the NFW dark matter density profile and the GNFW gas pressure profile.

Keyword arguments are redshift and overdensity.

Returns functions for gas temperature and gas mass density as a function of radius.

BayesJ.Model_Piecewise — Method

Model_Piecewise(radius, density, temperature, radius, density, temperature, ...)

Construct a piecewise model for the given data.

The model is constructed by linearly interpolating the given data. See LinearInterpolator for details on the interpolation behaviour.

Density is assumed to be in units of g/cm^3, temperature in units of keV, and radius in units of kpc.

BayesJ.Model_Vikhlinin2006 — Function

Model_Vikhlinin2006(n0::NumberDensity, n02::NumberDensity, rc::Unitful.Length, rc2::Unitful.Length, α, β, β2, ϵ, rs::Unitful.Length, T0::Unitful.Energy, TminT0, rcool::Unitful.Length, acool, rt::Unitful.Length, a, b, c; γ=3, kwargs...)

Generate a cluster profile using the highly free models from (Vikhlinin et al., 2006).

d was introduced to experiment with changing the behaviour of temperature at the core. Leave it at zero that match the original Vikhlinin equations.

BayesJ.Model_Vikhlinin2006 — Method

Model_Vikhlinin2006(n0::Real, n02::Real, rc::Real, rc2::Real, α, β, β2, ϵ, rs::Real, T0::Real, TminT0, rcool::Real, acool, rt::Real, a, b, c; γ=3, kwargs...)

Unitless wrapper for Model_Vikhlinin2006. Densities are in cm^-3 and radii are in kpc. Temperature is in keV.

BayesJ.prepare_model_mekal — Method

prepare_model_mekal(nHcol, energy_bins, z, [abundances])

Generate a function to calculate emission with MEKAL .

The returned function has signature

f!(flux::Vector{Float32}, T::Unitful.Energy, nH::Unitful.𝐋^-3)::Vector{Float32}

It works inplace on flux. After execution flux contains the volume emissivity in photons per cubic metre per second per bin of a source of the specified temperature and hydrogen number density. Photon redshift and time dilation effects based on the redshift z. It also applies the effects of absorption from passage through an area with a hydrogen column density nHcol.

Metal abundances are set by abundances. They include H, He, C, N, O, Ne, Na, Mg, Al, Si, S, Ar, Ca, Fe and Ni and are specified with respect to solar values. MEKAL uses the solar values from (Anders and Grevesse, 1989).

BayesJ.prepare_model_mekal_interpolation — Method

preparemodelmekalinterpolation(nHcol, energybins, z, [abundances, temperatures, hydrogen_densities])

Generate an interpolation function to calculate emission with MEKAL.

This wraps prepare_model_mekal to generate an interpolation function over the model. Interpolation points are determined from temperatures and hydrogen_densities.

Interpolation has a significant performance improvement over calling the model directly. No redshift is currently applied to energy bins - they should be assumed to be in the source frame. It may be wise to adjust this but then surface_brightness will need adjustment so it doesn't apply the correction twice.

BayesJ.priorcheck — Method

priorcheck(condition, likelihood)

Check that a condition holds and fallback to the given likelihood if not.

Simply call this function in your model. If the condition fails a PriorError is raised.

The likelihood is encapsulated in the PriorError so that a try-catch block in the function calling the model can extract it and return it as the likelihood generated by that prior set without going through observation generation and likelihood calculation.

The faux likelihood should be chosen to be sufficiently small (e.g. -1e100) that it will not be competing with the likelihood from good parameters. It should also be designed with a slope to point the sampler in the right direction. For instance if the constraint is that $a < b$ then $-1e100 * (1 + (a - b))$ would be good. This constraint ensures the likelihood grows larger when a and b grow closer together. The additional one ensures that $a=b$ won't set the likelihood to zero.

See the ultranest documentation for additional details.

BayesJ.sample — Method

sample(
    data::Dataset,
    energy_limits::NTuple{2, Unitful.Energy},
    cluster_model::Function,
    priors::AbstractVector{Prior},
    nhCol::SurfaceDensity,
    redshift::Real,
    x::NTuple{2,<:Real},
    y::NTuple{2,<:Real};
    bin_size::Real=10,
    use_interpolation::Bool=false,
    centre_radius::Int=0,
    mask=nothing,
    cache_size::Int64=1000000000
    )

Run Bayesian inference on a given set of data considering only the selected energy range.

The cluster model can be any function that takes a set of parameters matching the priors and a z keyword argument,

and returns two functions for the gas temperature and gas mass density as a function of radius.

The first two priors should always be named "x0" and "y0", giving cluster centre position ("x" and "y" are also accepted).

x and y are tuples of (min, max) in pixels. These crop the observation. Bin size is specified in the same units.

mask is optional. If included it should be a string pointing to a mask file using CIAO syntax. Only ellipses are supported.

centre_radius excludes some radius, in pixels, around the centre from analysis

Additional kwargs will be passed through to the next sample function.

BayesJ.sample — Method

sample(observed,observed_background, response_function, transform, obs_exposure_time, bg_exposure_time, redshift; prior_names, cluster_model, emission_model, param_wrapper, pixel_edge_angle)

Configure some necessary variables and launch ultranest.

The observed array includes the background. The first dimension is energy, the other two are spatial.

The response function includes both the RMF and ARF, as described in apply_response_function.

The emission model should be a function compatible with the requirements of the surface_brightness function, which it will be passed to.

The pixel edge angle describes the angular size observed by a single pixel in units such as arcseconds.

This area is assumed to be square with the edge angle giving the side length.

The first two priors should always be "x0" and "y0", giving cluster centre position and the order of prior_names must match the transform.

The cluster model should be a function that takes the parameters (and redshift as a kwarg) and returns (gas_temperature,gas_density) as functions of

radius which return their respective quantities with units.

param_wrapper takes the output of the transform function and adds any additional arguements necessary for the model.

centre_radius, mask and integration_limit are passed through to make_observation